# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Social App API",
    "version": "1.0.0",
    "description": "Endpoints for user registration, authentication, friend management, and location updates."
  },
  "servers": [
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/signup": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signup result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Authenticate a user and receive a session cookie",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful – session cookie set",
            "headers": {
              "Set-Cookie": {
                "description": "Session cookie",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/authCheck": {
        "get": {
            "summary": "Check if the user is authenticated and receive user info in that case",
            "security": [
                {
                    "cookieAuth": []
                }
            ],
            "responses": {
                "200": {
                    "description": "User is authenticated",
                    "content": {
                        "application/json": {
                            "schema": {
                               "$ref": "#/components/schemas/AuthCheckResponse"
                            }
                        }
                    }
                },
                "401": {
                    "description": "User is not authenticated"
                }
            }
        }
    },
    "/logout": {
        "post": {
            "summary": "Logout the authenticated user",
            "security": [
                {
                    "cookieAuth": []
                }
            ],
            "responses": {
                "200": {
                    "description": "Logout successful – session cookie cleared",
                    "headers": {
                        "Set-Cookie": {
                            "description": "Cleared session cookie",
                            "schema": {
                                "type": "string"
                            }
                        }
                    },
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenericSuccess"
                            }
                        }
                    }
                }
            }
        }
    },
    "/friend-requests": {
      "get": {
        "summary": "Get frend requests for the authenticated user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of friend requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendRequests"
                }
              }
            }
          }
        }
      }
    },
    "/friend-request/create": {
      "post": {
        "summary": "Send a friend request",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FriendRequestCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          }
        }
      }
    },
    "/friend-request/{id}/accept": {
      "post": {
        "summary": "Accept a pending friend request",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Friend request identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Friend added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "404": {
            "description": "Request not found"
          }
        }
      }
    },
    "/friend-request/{id}/reject": {
      "post": {
        "summary": "Reject a pending friend request",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Friend request identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "404": {
            "description": "Request not found"
          }
        }
      }
    },

    "/friend-request/{id}/cancel": {
      "post": {
        "summary": "Cancel a pending friend request",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Friend request identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Request canceled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "404": {
            "description": "Request not found"
          }
        }
      }
    },
    "/location": {
      "post": {
        "summary": "Report current position and retrieve recent friend locations",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationPush"
              }
            }
          }
        },
          "/register-token": {
            "post": {
              "summary": "Register an FCM device token for the current user",
              "security": [ { "cookieAuth": [] } ],
              "requestBody": {
                "required": true,
                "content": {
                  "application/json": {
                    "schema": { "$ref": "#/components/schemas/RegisterToken" }
                  }
                }
              },
              "responses": {
                "200": { "description": "Token registered", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericSuccess" } } } }
              }
            }
          },
          "/notify-friends": {
            "post": {
              "summary": "Send a push notification to a list of friends telling them you are nearby",
              "security": [ { "cookieAuth": [] } ],
              "requestBody": {
                "required": true,
                "content": {
                  "application/json": {
                    "schema": { "$ref": "#/components/schemas/CloseFriendsBody" }
                  }
                }
              },
              "responses": {
                "200": { "description": "Notification send result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericSuccess" } } } }
              }
            }
          },
        "responses": {
          "200": {
            "description": "Array of recent location entries for the supplied friends",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/LocationEntry" }
                }
              }
            }
          }
        }
      }
    },
    "/friends": {
      "get": {
        "summary": "List usernames of confirmed friends",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of friend usernames",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendList"
                }
              }
            }
          }
        }
      }
    },
    "/friends/{friend_username}": {
      "delete": {
        "summary": "Delete a friend connection",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "friend_username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Username of the friend to remove"
          }
        ],
        "responses": {
          "200": {
            "description": "Friend removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "404": {
            "description": "Friend not found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id"
      }
    },
    "schemas": {
      "SignupRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "example": "alice123" },
          "password": { "type": "string", "format": "password", "example": "P@ssw0rd!" }
        }
      },
      "SignupResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "detail": { "type": "string", "example": "Account created." }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "example": "alice123" },
          "password": { "type": "string", "format": "password", "example": "P@ssw0rd!" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "detail": { "type": "string", "example": "Logged in." }
        }
      },
      "AuthCheckResponse": {
          "type": "object",
          "required": ["username"],
          "properties": {
              "username": { "type": "string", "example": "alice123" }
          }
      },
      "LocationPush": {
        "type": "object",
        "required": ["latitude", "longitude"],
        "properties": {
          "latitude": { "type": "number", "format": "float", "example": 37.7749 },
          "longitude": { "type": "number", "format": "float", "example": -122.4194 }
        }
      },
      "FriendList": {
          "type": "array",
          "items": { "type": "string", "example": "bob456" }
      },
      "FriendRequestCreate": {
        "type": "object",
        "required": ["friendUsername"],
        "properties": {
            "friendUsername": { "type": "string", "example": "bob456" }
        }
      },
      "FriendRequest": {
          "type": "object",
          "required": ["id", "friendName"],
          "properties": {
              "id": { "type": "string" },
              "friendName": { "type": "string", "example": "bob456" }
          }
      },
      "FriendRequestsList": {
          "type": "array",
          "items": { "$ref": "#/components/schemas/FriendRequest" }
      },
      "FriendRequests": {
          "type": "object", 
          "required": ["requestsToYou", "requestsFromYou"],
          "properties": {
              "requestsToYou": { "$ref": "#/components/schemas/FriendRequestsList" },
              "requestsFromYou": { "$ref": "#/components/schemas/FriendRequestsList" }
          }
      },
      "GenericSuccess": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "detail": { "type": "string", "example": "Operation completed." }
        }
      },
      "LocationEntry": {
        "type": "object",
        "required": ["username", "latitude", "longitude", "ts"],
        "properties": {
          "username": { "type": "string" },
          "latitude": { "type": "number", "format": "float" },
          "longitude": { "type": "number", "format": "float" },
          "ts": { "type": "string", "format": "date-time" }
        }
      }
      ,
      "RegisterToken": {
        "type": "object",
        "required": ["token"],
        "properties": {
          "token": { "type": "string", "example": "fcm_device_token_here" }
        }
      },
      "CloseFriendsBody": {
        "type": "object",
        "required": ["friends"],
        "properties": {
          "friends": { "type": "array", "items": { "type": "string" }, "example": ["bob"] }
        }
      }
    }
  }
}